+++ src/integrationTest/java/com/mongodb/kafka/connect/MongoSourceConnectorIntegrationTest.java
@@ -27,14 +27,11 @@ import static java.util.Collections.singletonList;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.IntStream.rangeClosed;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertIterableEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Properties;
 import java.util.stream.StreamSupport;
 
@@ -47,7 +44,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import org.bson.BsonDocument;
 import org.bson.Document;
 
 import com.mongodb.client.MongoCollection;
@@ -294,43 +290,46 @@ public class MongoSourceConnectorIntegrationTest extends MongoKafkaTestCase {
           MongoSourceConfig.COLLECTION_CONFIG, coll.getNamespace().getCollectionName());
       sourceProperties.put(MongoSourceConfig.HEARTBEAT_INTERVAL_MS_CONFIG, "1000");
       sourceProperties.put(MongoSourceConfig.HEARTBEAT_TOPIC_NAME_CONFIG, heartbeatTopic);
-      sourceProperties.put("offset.flush.interval.ms", "1000");
 
       addSourceConnector(sourceProperties);
 
-      String heartBeat =
-          getProducedStrings(heartbeatTopic, 5).stream()
-              .reduce((first, second) -> second)
-              .orElse("");
-      assertFalse(heartBeat.isEmpty());
-
       insertMany(rangeClosed(1, 50), coll);
-      List<String> producedKeys = getProducedKeys(coll, 50);
-      BsonDocument lastSeenKey =
-          BsonDocument.parse(
-                  producedKeys.stream()
-                      .reduce((first, second) -> second)
-                      .orElseGet(() -> "{_id: {}}"))
-              .getDocument("_id");
-
       insertMany(rangeClosed(1, 50), altColl);
-      sleep();
-
-      heartBeat =
-          getProducedStrings(heartbeatTopic, 5).stream()
-              .reduce((first, second) -> second)
-              .orElse("");
-      assertFalse(heartBeat.isEmpty());
-      assertNotEquals(lastSeenKey, BsonDocument.parse(heartBeat));
+      getProducedStrings(heartbeatTopic, 1);
 
       stopStartSourceConnector(sourceProperties);
 
-      String expectedResumeString =
-          "Resuming the change stream after the previous offset: " + heartBeat;
-      assertTrue(
+      boolean resumedFromHeartbeat =
           logCapture.getEvents().stream()
               .map(e -> e.getMessage().toString())
-              .anyMatch(e -> Objects.equals(e, expectedResumeString)));
+              .anyMatch(e -> e.startsWith("Resume token from heartbeat"));
+
+      assertTrue(resumedFromHeartbeat);
+    }
+  }
+
+  @Test
+  @DisplayName("Ensure Source provides friendly error messages for invalid pipelines")
+  void testSourceHasFriendlyErrorMessagesForInvalidPipelines() {
+    assumeTrue(isGreaterThanFourDotZero());
+    try (LogCapture logCapture = new LogCapture(Logger.getLogger(MongoSourceTask.class))) {
+      MongoCollection<Document> coll = getAndCreateCollection();
+
+      Properties sourceProperties = new Properties();
+      sourceProperties.put(
+          MongoSourceConfig.DATABASE_CONFIG, coll.getNamespace().getDatabaseName());
+      sourceProperties.put(
+          MongoSourceConfig.COLLECTION_CONFIG, coll.getNamespace().getCollectionName());
+      sourceProperties.put(PIPELINE_CONFIG, "[{'$group': {_id: 1 }}]");
+
+      addSourceConnector(sourceProperties);
+
+      boolean containsIllegalChangeStreamOperation =
+          logCapture.getEvents().stream()
+              .map(e -> e.getMessage().toString())
+              .anyMatch(e -> e.startsWith("Illegal $changeStream operation"));
+
+      assertTrue(containsIllegalChangeStreamOperation);
     }
   }